refactor: ExceptionResponse 클래스를 exception 디렉토리에 패키징
