+++ src/main/java/com/backendoori/ootw/PostController.java
@@ -3,10 +3,11 @@ package com.backendoori.ootw;
 import java.net.URI;
 import java.util.List;
 import java.util.NoSuchElementException;
-import com.backendoori.ootw.ExceptionResponse.FieldErrorDetail;
 import com.backendoori.ootw.dto.PostDetailInfo;
 import com.backendoori.ootw.dto.PostSaveRequest;
 import com.backendoori.ootw.dto.PostSaveResponse;
+import com.backendoori.ootw.exception.ExceptionResponse;
+import com.backendoori.ootw.exception.ExceptionResponse.FieldErrorDetail;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
